requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
torch==2.2.2
numpy==1.26.4
pycryptodome==3.20.0
